requirements.txt
# Requirements for torch.py

# Workflow engine
prefect
SqlAlchemy
docker-compose # for Prefect Server backend

# Web
Flask
Flask_Mail
Flask_Migrate
Flask_SQLAlchemy
Flask-Security-Too
flask-socketio
WTForms
email_validator

# TORCH Tasks
pysftp # SFTP upload
boto3 # Amazon S3 upload
pillow # Derivative image processing

pyzbar
tqdm